added logic for handling hostname and port
